LPD-56439 Refactor chart
