+++ modules/dxp/apps/analytics/analytics-reports-js-components-web/src/main/resources/META-INF/resources/js/InContextAnalytics.tsx
@@ -8,8 +8,7 @@ import React from 'react';
 import {ContextProvider} from './Context';
 import {CheckPermissions} from './components/cms/CheckPermissions';
 import {Metrics} from './components/cms/Metrics';
-import {MetricsChart} from './components/cms/MetricsChart';
-import {ChartsAndTableView} from './components/cms/TableAndChartsView';
+import {CurrentVsPrevious} from './components/cms/current-vs-previous/CurrentVsPrevious';
 
 import '../css/in-context-analytics.scss';
 import {AssetTypes} from './types/global';
@@ -31,10 +30,7 @@ const InContextAnalytics: React.FC<IInContextAnalyticsProps> = ({assetId}) => {
 				>
 					<Metrics />
 
-					<ChartsAndTableView
-						chartsView={<MetricsChart />}
-						tableView={<>Table!</>}
-					/>
+					<CurrentVsPrevious />
 				</ContextProvider>
 			</CheckPermissions>
 		</div>
